+++ WLTerminal.m
@@ -139,24 +139,24 @@
 
 # pragma mark -
 # pragma mark Access Data
-- (attribute)attrAtRow:(int)r 
-				column:(int)c {
+- (attribute)attrAtRow:(NSInteger)r
+				column:(NSInteger)c {
 	return _grid[r][c].attr;
 }
 
-- (NSString *)stringAtIndex:(int)begin 
-					 length:(int)length {
-    int i, j;
+- (NSString *)stringAtIndex:(NSInteger)begin
+					 length:(NSInteger)length {
+    NSInteger i, j;
     //unichar textBuf[length + 1];
     unichar firstByte = 0;
-    int bufLength = 0;
-    int spacebuf = 0;
+    NSInteger bufLength = 0;
+    NSInteger spacebuf = 0;
 	if (begin + length > _maxRow * _maxColumn) {
 		length = _maxRow * _maxColumn - begin;
 	}
     for (i = begin; i < begin + length; i++) {
-        int x = i % _maxColumn;
-        int y = i / _maxColumn;
+        NSInteger x = i % _maxColumn;
+        NSInteger y = i / _maxColumn;
         if (x == 0 && i != begin && i - 1 < begin + length) { // newline
 			// REVIEW: why we need to update double byte state?????
             [self updateDoubleByteStateForRow:y];
@@ -237,17 +237,17 @@
 	return [self stringAtIndex:row * _maxColumn length:_maxColumn];
 }
 
-- (cell *)cellsOfRow:(int)r {
+- (cell *)cellsOfRow:(NSInteger)r {
 	return _grid[r];
 }
 
-- (cell)cellAtIndex:(int)index {
+- (cell)cellAtIndex:(NSInteger)index {
 	return _grid[index / _maxColumn][index % _maxColumn];
 }
 
 # pragma mark -
 # pragma mark Update State
-- (void)updateDoubleByteStateForRow:(int)r {
+- (void)updateDoubleByteStateForRow:(NSInteger)r {
 	cell *currRow = _grid[r];
 	int db = 0;
 	BOOL isDirty = NO;
